+++ examples/common/TransportAdapter.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "I2cTransport.h"
+
+namespace transport_adapter {
+using namespace transport;
+}